add yosys file for maxpool

+++ Benchmarks2/maxpool/compile.sh
@@ -1,5 +1,16 @@
-#!/bin/bash
+echo "Use -d to compile with Synopsys Design Compiler"
+
 mkdir -p syn
 
-design_vision -no_gui -f maxpool.dcsh
-rm *.pvl *.syn *.mr *.log *.svf
+if [ $1 ] && [ $1 = "-d" ]; then
+	design_vision -no_gui -f maxpool.dcsh
+	rm *.pvl *.syn *.mr *.log *.svf
+else
+	yosys -c maxpool.tcl
+fi
+
+for verilogfile in syn/*.v
+do
+  empfile=${verilogfile%.*}.emp
+  ../../Verilog2EMP/bin/V2EMP_Main -i $verilogfile -o $empfile --log2std 
+done
